Refactor into a workspace with two packages
The jmdict_to_html folder contains our existing code. The html_dictionary_previewer is a rudimentary static site for previewing the generated dictionaries. It contains some code and stylesheets adapted from a previous attempt at this project, which need to be cleaned up and improved to apply to the current round of work.